requirements.txt
flask
pandas
numpy
scikit-learn
google-genai